Finalização exercício apis-labenusers
