+++ modulo3/apis-labenusers/apis-labenusers/src/components/TelaDois.js
@@ -1,8 +1,15 @@
 import React from "react";
 
-const TelaDois = ({users}) =>{
+const TelaDois = ({users, setTrocarTela}) =>{
+
+    const trocar = (e)=>{
+        e.preventDefault()
+        setTrocarTela(false)
+    } 
+
     return(
         <div>
+            <button onClick={trocar}>Trocar Tela</button>        
             {users}
         </div>
     )
